# backend/laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Byspolye Api Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://byspolye.api.skillzserver.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/student/login": {
            "post": {
                "tags": [
                    "Student"
                ],
                "summary": "Student Login",
                "operationId": "10bb6cce779797b42ad7d7945a452693",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Student login data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Student loggedin successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "user logged in"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "first_name": {
                                                    "type": "string"
                                                },
                                                "middle_name": {
                                                    "type": "string"
                                                },
                                                "last_name": {
                                                    "type": "string"
                                                },
                                                "maiden_name": {
                                                    "type": "string"
                                                },
                                                "reg_number": {
                                                    "type": "string"
                                                },
                                                "dob": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "email_verified_at": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "password": {
                                                    "type": "string"
                                                },
                                                "country_id": {
                                                    "type": "integer"
                                                },
                                                "state_id": {
                                                    "type": "integer"
                                                },
                                                "lga_id": {
                                                    "type": "integer"
                                                },
                                                "token": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "integer",
                                                    "example": "1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}